rename render to renderFullscreen
