+++ packages/core/src/index.ts
@@ -2,7 +2,7 @@ export { terminal } from 'terminal-kit';
 
 export { default as renderToJson } from './renderToJson';
 export { default as renderToString } from './renderToString';
-export { default as render, exit } from './render';
+export { default as renderFullscreen, exit } from './renderFullscreen';
 
 export { default as View, Props as ViewProps } from './components/View';
 export { default as Text, Props as TextProps } from './components/Text';
