+++ packages/core/src/renderFullscreen.ts
